django/requirements.txt
Django==4.1.7
psycopg2==2.9.5
djangorestframework==3.14.0
requests==2.28.2